战前数据接入；删除战前卡片对比；适配新Ui;
git-svn-id: svn://ci.ftxgame.local:8199/repos/ArmChairQuatebackDoc/sourcecode/arm_chair_quaterback@647324 11eda05a-b276-4d74-a309-1d70aa31e1c1

+++ lib/common/entities/team_info_entity.dart
@@ -20,7 +20,7 @@ class TeamInfoEntity {
   late int powerReplyTime;
   late int currentGuessWinningStreak;
   late int currentGameWinningStreak;
-  late int playerReadiness;
+  late double playerReadiness;
   late int accountId;
   late int firstLoginTime;
   late int secretaryId;
